Modelos y primera actividad

+++ database/migrations/2022_04_06_153332_create_zapatos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('zapatos', function (Blueprint $table) {
             $table->id();
-            $table->integer('codigo')->unique();
+            $table->bigInteger('codigo')->unique();
             $table->string('denominacion');
             $table->decimal('precio', 6, 2);
             $table->timestamps();
